Backend/requirements.txt
Flask
Flask-Cors
google-generativeai
PyPDF2
textract